+++ core-ng/src/main/java/core/framework/internal/log/appender/KafkaAppender.java
@@ -45,14 +45,14 @@ public final class KafkaAppender implements LogAppender {
         statMapper = new JSONMapper<>(StatMessage.class);
         try {
             Map<String, Object> config = Map.of(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, uri,
-                    ProducerConfig.ACKS_CONFIG, "0",                                                        // no acknowledge to maximize performance
-                    ProducerConfig.CLIENT_ID_CONFIG, "log-forwarder",                                       // if not specify, kafka uses producer-${seq} name, also impact jmx naming
+                    ProducerConfig.ACKS_CONFIG, "0",                                        // no acknowledge to maximize performance
+                    ProducerConfig.CLIENT_ID_CONFIG, "log-forwarder",                       // if not specify, kafka uses producer-${seq} name, also impact jmx naming
                     ProducerConfig.COMPRESSION_TYPE_CONFIG, CompressionType.SNAPPY.name,
-                    ProducerConfig.DELIVERY_TIMEOUT_MS_CONFIG, (int) Duration.ofSeconds(60).toMillis(),     // DELIVERY_TIMEOUT_MS_CONFIG is INT type
-                    ProducerConfig.LINGER_MS_CONFIG, 50,
-                    ProducerConfig.RECONNECT_BACKOFF_MS_CONFIG, Duration.ofMillis(500).toMillis(),          // longer backoff to reduce cpu usage when kafka is not available
-                    ProducerConfig.RECONNECT_BACKOFF_MAX_MS_CONFIG, Duration.ofSeconds(5).toMillis(),
-                    ProducerConfig.MAX_BLOCK_MS_CONFIG, Duration.ofSeconds(30).toMillis());                 // metadata update timeout, shorter than default, to get exception sooner if kafka is not available
+                    ProducerConfig.DELIVERY_TIMEOUT_MS_CONFIG, 60 * 1000,                   // 60s, type is INT
+                    ProducerConfig.LINGER_MS_CONFIG, 50L,
+                    ProducerConfig.RECONNECT_BACKOFF_MS_CONFIG, 500L,                        // longer backoff to reduce cpu usage when kafka is not available
+                    ProducerConfig.RECONNECT_BACKOFF_MAX_MS_CONFIG, 5L * 1000,               // 5s
+                    ProducerConfig.MAX_BLOCK_MS_CONFIG, 30L * 1000);                         // 30s, metadata update timeout, shorter than default, to get exception sooner if kafka is not available
             var serializer = new ByteArraySerializer();
             producer = new KafkaProducer<>(config, serializer, serializer);
             producerMetrics.set(producer.metrics());